processing/flare_capa/requirements.txt
flare-capa==5.1.0
